Add scan type generation and filter validators (#40)
За время написания кода уже и забыл что тут было, но вот что-то да вышло
в итоге.

Есть один класс который хочется потом засурсгенить
FilterValidatorApplier, но чтобы не растягивать это будет в отдельном
пре

+++ Issueneter.Persistence/ScanStorage.cs
@@ -39,7 +39,7 @@ public class ScanStorage
 
         var @params = new
         {
-            type = (int)creation.Type,
+            type = creation.Type,
             acc = creation.Owner,
             repo = creation.Repo,
             chat = creation.ChatId,
